Rewrite AreaLoader
