+++ packages/bento-design-system/src/index.ts
@@ -5,7 +5,7 @@ import * as illustrations from "./Illustrations";
 import * as defaultConfigs from "./util/defaultConfigs";
 
 export * from "./Actions/Actions";
-export * from "./AreaLoader/createAreaLoader";
+export * from "./AreaLoader/AreaLoader";
 export * from "./Avatar/createAvatar";
 export * from "./Banner/Banner";
 export * from "./BentoProvider";
